requirements.txt
Flask==3.0.0
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0
python-dotenv==1.0.0
